+++ app/src/main/java/com/example/customermanagementsystem/fragments/GroupStudentsFragment.kt
@@ -16,14 +16,13 @@ import com.example.customermanagementsystem.R
 import com.example.customermanagementsystem.repository.ViewModel
 import com.example.customermanagementsystem.repository.ViewModelFactory
 import com.example.customermanagementsystem.adapter.GroupStudentsAdapter
-import com.example.customermanagementsystem.models.GroupDTO
 import com.example.customermanagementsystem.repository.Repository
 import kotlinx.android.synthetic.main.fragment_group_students.*
 
 class GroupStudentsFragment : Fragment() {
 
     private lateinit var viewModel: ViewModel
-    private var groupId:Long = 0
+    private var groupId:Int = 0
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
         val view = inflater.inflate(R.layout.fragment_group_students, container, false)
@@ -32,7 +31,7 @@ class GroupStudentsFragment : Fragment() {
 
     override fun onAttach(context: Context) {
         super.onAttach(context)
-        groupId = (parentFragment as WrapGroupDataFragment).groupId
+        groupId = (parentFragment as WrapGroupDataFragment).groupId.toInt()
         Log.e("groupStudentsID", groupId.toString())
     }
 
@@ -68,7 +67,7 @@ class GroupStudentsFragment : Fragment() {
 
         fab_add_student.setOnClickListener {
             var bundle = Bundle()
-            bundle.putLong("groupId", groupId)
+            bundle.putInt("groupId", groupId)
             findNavController().navigate(R.id.action_wrapGroupDataFragment_to_groupAddStudentFragment, bundle)
         }
     }
